+++ V2/Cargohub/controllers/itemtypecontroller.cs
@@ -24,6 +24,15 @@ public class ItemTypeController : ControllerBase
     [HttpGet()]
     public ActionResult<IEnumerable<ItemCS>> GetAllItemtypes()
     {
+        List<string> listOfAllowedRoles = new List<string>() { "Admin", "Warehouse Manager", "Inventory Manager",
+                                                               "Analyst", "Logistics", "Sales" };
+        var userRole = HttpContext.Items["UserRole"]?.ToString();
+
+        if (userRole == null || !listOfAllowedRoles.Contains(userRole))
+        {
+            return Unauthorized();
+        }
+
         var itemtype = _itemtypeService.GetAllItemtypes();
         return Ok(itemtype);
     }
@@ -32,6 +41,15 @@ public class ItemTypeController : ControllerBase
     [HttpGet("{id}/items")]
     public ActionResult<IEnumerable<ItemCS>> GetAllItemsInItemType(int id)
     {
+        List<string> listOfAllowedRoles = new List<string>() { "Admin", "Warehouse Manager", "Inventory Manager",
+                                                               "Analyst", "Logistics", "Sales" };
+        var userRole = HttpContext.Items["UserRole"]?.ToString();
+
+        if (userRole == null || !listOfAllowedRoles.Contains(userRole))
+        {
+            return Unauthorized();
+        }
+
         var itemsInItemType = _itemService.GetAllItemsInItemType(id);
         if (itemsInItemType is null)
         {
@@ -44,6 +62,15 @@ public class ItemTypeController : ControllerBase
     [HttpGet("{id}")]
     public ActionResult<ItemTypeCS> GetItemById(int id)
     {
+        List<string> listOfAllowedRoles = new List<string>() { "Admin", "Warehouse Manager", "Inventory Manager",
+                                                               "Analyst", "Logistics", "Sales" };
+        var userRole = HttpContext.Items["UserRole"]?.ToString();
+
+        if (userRole == null || !listOfAllowedRoles.Contains(userRole))
+        {
+            return Unauthorized();
+        }
+
         var itemtype = _itemtypeService.GetItemById(id);
         if (itemtype == null)
         {
@@ -56,18 +83,55 @@ public class ItemTypeController : ControllerBase
     [HttpPost()]
     public async Task<IActionResult> CreateItemType([FromBody] ItemTypeCS itemtype)
     {
+        List<string> listOfAllowedRoles = new List<string>() { "Admin", "Warehouse Manager", "Inventory Manager" };
+        var userRole = HttpContext.Items["UserRole"]?.ToString();
+
+        if (userRole == null || !listOfAllowedRoles.Contains(userRole))
+        {
+            return Unauthorized();
+        }
+
         if (itemtype == null)
         {
             return BadRequest("ItemGroup cannot be null");
         }
 
-        var createditemtype = await _itemtypeService.CreateItemType(itemtype);
+        var createditemtype = _itemtypeService.CreateItemType(itemtype);
         return CreatedAtAction(nameof(GetItemById), new { id = createditemtype.Id }, createditemtype);
+    }
+
+    // POST: /itemtypes/multiple
+    [HttpPost("multiple")]
+    public ActionResult<ItemTypeCS> CreateMultipleItemTypes([FromBody] List<ItemTypeCS> newItemTypes)
+    {
+        List<string> listOfAllowedRoles = new List<string>() { "Admin", "Warehouse Manager", "Sales", "Logistics" };
+        var userRole = HttpContext.Items["UserRole"]?.ToString();
+
+        if (userRole == null || !listOfAllowedRoles.Contains(userRole))
+        {
+            return Unauthorized();
+        }
+
+        if (newItemTypes is null)
+        {
+            return BadRequest("ItemType data is null");
+        }
+
+        var createdItemTypes = _itemtypeService.CreateMultipleItemTypes(newItemTypes);
+        return StatusCode(StatusCodes.Status201Created, createdItemTypes);
     }
 
     [HttpPut("{id}")]
     public async Task<ActionResult<ItemTypeCS>> UpdateItemType(int id, [FromBody] ItemTypeCS itemType)
     {
+        List<string> listOfAllowedRoles = new List<string>() { "Admin", "Warehouse Manager", "Inventory Manager" };
+        var userRole = HttpContext.Items["UserRole"]?.ToString();
+
+        if (userRole == null || !listOfAllowedRoles.Contains(userRole))
+        {
+            return Unauthorized();
+        }
+
         if (id != itemType.Id)
         {
             return BadRequest();
@@ -96,6 +160,14 @@ public class ItemTypeController : ControllerBase
     [HttpDelete("{id}")]
     public ActionResult DeleteItemType(int id)
     {
+        List<string> listOfAllowedRoles = new List<string>() { "Admin" };
+        var userRole = HttpContext.Items["UserRole"]?.ToString();
+
+        if (userRole == null || !listOfAllowedRoles.Contains(userRole))
+        {
+            return Unauthorized();
+        }
+
         var itemtype = _itemtypeService.GetItemById(id);
         if (itemtype == null)
         {
@@ -104,9 +176,20 @@ public class ItemTypeController : ControllerBase
         _itemtypeService.DeleteItemType(id);
         return Ok();
     }
+
     [HttpDelete("batch")]
-    public ActionResult DeleteItemTypes(List<int> ids){
-        if(ids is null){
+    public ActionResult DeleteItemTypes(List<int> ids)
+    {
+        List<string> listOfAllowedRoles = new List<string>() { "Admin", "Warehouse Manager", "Inventory Manager" };
+        var userRole = HttpContext.Items["UserRole"]?.ToString();
+
+        if (userRole == null || !listOfAllowedRoles.Contains(userRole))
+        {
+            return Unauthorized();
+        }
+
+        if (ids is null)
+        {
             return NotFound();
         }
         _itemtypeService.DeleteItemTypes(ids);
